// src/services/types/audit-log/audit-log-search-audit-logs.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

import type {ApiResultOnePageAuditLogPageResponse} from "../api-result-one-page-audit-log-page-response.ts";

export type AuditLogSearchAuditLogsQueryParams = {
	/**
	 * @description 관리자 아이디
	 * @type string | undefined
	 */
	adminId?: string;
	/**
	 * @description 타임스탬프 시작
	 * @type string | undefined
	 */
	timestampFrom?: string;
	/**
	 * @description 타임스탬프 종료
	 * @type string | undefined
	 */
	timestampTo?: string;
	/**
	 * @description 조회 대상 필드 목록
	 * @type array | undefined
	 */
	field?: string[];
	/**
	 * @description 조회 연산자 목록
	 * @type array | undefined
	 */
	operator?: string[];
	/**
	 * @description 조회 값 목록
	 * @type array | undefined
	 */
	value?: string[];
	/**
	 * @type integer | undefined, int32
	 */
	page?: number;
	/**
	 * @type integer | undefined, int32
	 */
	size?: number;
};

/**
 * @description OK
 */
export type AuditLogSearchAuditLogs200 = ApiResultOnePageAuditLogPageResponse;

export type AuditLogSearchAuditLogsQueryResponse = AuditLogSearchAuditLogs200;

export type AuditLogSearchAuditLogsQuery = {
	Response: AuditLogSearchAuditLogs200;
	QueryParams: AuditLogSearchAuditLogsQueryParams;
	Errors: any;
};