// src/services/types/audit-log-page-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export type AuditLogPageResponse = {
	/**
	 * @description 감사 로그 아이디
	 * @type integer, int64
	 */
	id: number;
	/**
	 * @description 관리자 아이디
	 * @type string | undefined
	 */
	adminId?: string;
	/**
	 * @description 관리자 이름
	 * @type string | undefined
	 */
	adminName?: string;
	/**
	 * @description 설명
	 * @type string | undefined
	 */
	description?: string;
	/**
	 * @description 요청 URL
	 * @type string | undefined
	 */
	url?: string;
	/**
	 * @description 요청 메서드
	 * @type string | undefined
	 */
	method?: string;
	/**
	 * @description 응답 상태
	 * @type string | undefined
	 */
	status?: string;
	/**
	 * @description 요청 본문
	 * @type string | undefined
	 */
	requestBody?: string;
	/**
	 * @description 실행 시간(ms)
	 * @type integer | undefined, int32
	 */
	executionTimeMs?: number;
	/**
	 * @description 요청 시작 시각
	 * @type string | undefined
	 */
	start?: string;
	/**
	 * @description 요청 종료 시각
	 * @type string | undefined
	 */
	end?: string;
	/**
	 * @description 로그 타임스탬프
	 * @type string | undefined
	 */
	timestamp?: string;
	/**
	 * @description 생성 일시
	 * @type string, date-time
	 */
	createdAt: string;
};